# ~9x performance improvement in screen_grab and resize