Added window title search.
Keep win32 DC objects for re-use.
Moved resizing to win32 api.
Removed cv2 dependendancy from grabscreen.

Benchmark:
cv2 version
	%timeit screen = grab_screen(region=(0,40,1920,1120));screen = cv2.resize(screen, (480,270));screen = cv2.cvtColor(screen, cv2.COLOR_BGR2RGB)
	3500 µs ± 1230 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

win32 version
	%timeit screen = grab_screen(window_title='Spyder',region=(0,40,1920,1120),scale=0.25)
	388 µs ± 15.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

Thanks for the awesome project and in-depth explanations and code.
-Dustin (aka SaintD_1970)