@Sentdex 2017-06-30
Hmm, I get an attribute error

AttributeError: 'windows_screen_grab' object has no attribute '_hwnd'

I am seeing failure on the if search_str in getwindowtxt line, that is never True for me. 

For my window title substring I have 'Grand Theft Auto V'