@dustinandrews 2017-06-30
Yes, annoyingly I didn't test it very well! Sorry about that. I got the same error. Will make an update later. It's a problem with a missing _ on the first hwnd in the class.

________________________________
From: Harrison <notifications@github.com>
Sent: Friday, June 30, 2017 5:57:18 AM
To: Sentdex/pygta5
Cc: Dustin Andrews; Author
Subject: Re: [Sentdex/pygta5] ~9x performance improvement in screen_grab and resize (#52)


Hmm, I get an attribute error

AttributeError: 'windows_screen_grab' object has no attribute '_hwnd'

I am seeing failure on the if search_str in getwindowtxt line, that is never True for me.

For my window title substring I have 'Grand Theft Auto V'

—
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://github.com/Sentdex/pygta5/pull/52#issuecomment-312260173>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AAtkv8uVOvJW2tMgUNV5X4oHY89slYhaks5sJPCugaJpZM4OGAmX>.