@dustinandrews 2017-07-01
Tested working now. Thank you for your patience. I wrote out data and was able to read it back and render it as an image. FYI: You can just pass a snip of the window title. Like 'Grand'.