@Sentdex 2017-07-01
Very cool, so NOW, how does one use this with open cv? I got it to find the window, and take a screenshot. Shape is (3, 270, 480), but I cannot do an cv2.imshow on it, since the channels comes first, not last. Okay, no problem, so next I do a 

screen = np.reshape(screen, (480, 270, 3))

... but now the image is all-white :(  

Ideas?