@dustinandrews 2017-07-06
I also realized you may want to use np.swapaxes rather than reshape.


Try this instead:


channels_last = screenshot.swapaxes(0,2).swapaxes(0,1)

________________________________
From: Dustin Andrews <dustinandrews@hotmail.com>
Sent: Thursday, July 6, 2017 1:56:06 PM
To: Sentdex/pygta5
Subject: Re: [Sentdex/pygta5] ~9x performance improvement in screen_grab and resize (#52)


Sorry for the slow response. I took a computer free break over the fourth. Hmm, I was able to view the channels with pyplots imshow when I tested it.


I've made some fixes and pushed my changes. Can you test with a command window just as a common baseline? I don't have GTA, but it's working with another game.

Some things I noticed:


  *   If you have the window minimized when you start the script it will get the wrong size and co-ordinates.

  *   If your game is full screen, I am not sure what will happen
  *   Some programs don't screenshot correctly, Chrome and Outlook for example. Chrome is just empty data and Outlook is missing a lot of data.
  *   Title bars sometimes don't show on screen the same as they report to windows. I put a "print" in the window handle enumeration to see them all.

I added a parameter for "window_handle" so you can use your same time delay strategy and just pass the window handle rather than a title. I hope that helps.



________________________________
From: Harrison <notifications@github.com>
Sent: Saturday, July 1, 2017 6:46:30 AM
To: Sentdex/pygta5
Cc: Dustin Andrews; Author
Subject: Re: [Sentdex/pygta5] ~9x performance improvement in screen_grab and resize (#52)


Very cool, so NOW, how does one use this with open cv? I got it to find the window, and take a screenshot. Shape is (3, 270, 480), but I cannot do an cv2.imshow on it, since the channels comes first, not last. Okay, no problem, so next I do a

screen = np.reshape(screen, (480, 270, 3))

... but now the image is all-white :(

Ideas?

—
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://github.com/Sentdex/pygta5/pull/52#issuecomment-312433067>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AAtkvyqNN_4zt1E65w6hbnJiMYqXrIaFks5sJk22gaJpZM4OGAmX>.