@Timotaam 2017-07-17
All you need to do if you want 'cv2' output is change the `get_rgb_from_bits` function to this:

`
    def get_rgb_from_bits(self, bits):
        """
        Reshape to rgb and strip the alpha channel.
        :parameter bits: numpy array in the format (r, g, b, a, r, g, b, a, ...)
        """    
        bits.shape = (self._dest_height,self._dest_width,4)
        self._rgb = bits
        return self._rgb
`

Though you might want to strip the alpha channel for size reasons? 

This is a really nice and fast solution, I like it