@dustinandrews 2017-07-18
I don't know what the Alpha channel would be good for in this use case so I just drop it to save memory and speed up saving to disk when I am done capturing.

________________________________
From: Timotaam <notifications@github.com>
Sent: Sunday, July 16, 2017 5:28:22 PM
To: Sentdex/pygta5
Cc: Dustin Andrews; Mention
Subject: Re: [Sentdex/pygta5] ~9x performance improvement in screen_grab and resize (#52)


All you need to do if you want 'cv2' output is change the get_rgb_from_bits function to this:

def get_rgb_from_bits(self, bits): """ Reshape to rgb and strip the alpha channel. :parameter bits: numpy array in the format (r, g, b, a, r, g, b, a, ...) """ bits.shape = (self._dest_height,self._dest_width,4) self._rgb = bits return self._rgb

Though you might want to strip the alpha channel for size reasons?

This is a really nice and fast solution, I like it

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub<https://github.com/Sentdex/pygta5/pull/52#issuecomment-315649162>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AAtkv2n-43m165enDLr0SRId-w_1egQmks5sOqqmgaJpZM4OGAmX>.